WIP modifying file structure
